feat #182 : 챌린지 상세 group id 포함 조회 네비게이션 설정, ID 타입 Long으로 변환

+++ android/feature/teacher/challenge/src/main/java/com/sixkids/teacher/challenge/history/ChallengeHistoryViewModel.kt
@@ -21,7 +21,7 @@ class ChallengeHistoryViewModel @Inject constructor(
 ) {
     var challengeHistory: Flow<PagingData<Challenge>>? = null
 
-    fun navigateChallengeDetail(challengeId: Int) = postSideEffect(
+    fun navigateChallengeDetail(challengeId: Long) = postSideEffect(
         ChallengeHistoryEffect.NavigateToChallengeDetail(challengeId)
     )
